Add OpenAI handler endpoint tests
