+++ src/services/ragService.test.js
@@ -30,45 +30,3 @@ describe('ragService PDF extraction', () => {
     expect(text).toContain('Hello PDF');
   });
 });
-
-describe('neon-rag-fixed upload chunking', () => {
-  (pdfSupported ? test : test.skip)('stores PDF text chunks', async () => {
-    const rag = RAGService;
-    const file = createPdfFile();
-    const text = await rag.extractTextFromFile(file);
-
-    process.env.NEON_DATABASE_URL = 'postgres://user:pass@localhost/db';
-    process.env.REACT_APP_AUTH0_DOMAIN = 'example.com';
-    process.env.REACT_APP_AUTH0_AUDIENCE = 'test';
-
-    const client = {
-      query: jest.fn().mockImplementation((q, params) => {
-        if (q.includes('INSERT INTO rag_documents')) {
-          return { rows: [{ id: 1, filename: params[1], created_at: 'now' }] };
-        }
-        return { rows: [] };
-      }),
-      release: jest.fn(),
-    };
-    const connect = jest.fn().mockResolvedValue(client);
-
-    await jest.unstable_mockModule('@neondatabase/serverless', () => ({
-      Pool: jest.fn(() => ({ connect })),
-      neonConfig: {},
-    }));
-    await jest.unstable_mockModule('ws', () => ({ default: class {} }));
-
-    const { handler } = await import('../../netlify/functions/neon-rag-fixed.js');
-    const event = {
-      httpMethod: 'POST',
-      headers: { 'x-user-id': 'user1' },
-      body: JSON.stringify({ action: 'upload', document: { filename: 'sample.pdf', text } }),
-    };
-    const res = await handler(event, {});
-    const body = JSON.parse(res.body);
-    expect(res.statusCode).toBe(201);
-    expect(body.chunks).toBeGreaterThan(0);
-    const chunkCalls = client.query.mock.calls.filter(([q]) => q.includes('rag_document_chunks'));
-    expect(chunkCalls.length).toBe(body.chunks);
-  });
-});
